Optimize startCommand
For every monitor in xmobar, we currently spawn two threads during
startup. But one of the threads from it get's exited as soon as it's
started because this is the default definition of trigger:

```
    trigger :: e -> (Maybe SignalType -> IO ()) -> IO ()
    trigger _ sh  = sh Nothing
```

So, for the `Nothing`, the expression `maybe (return ()) (atomically
. putTMVar sig)` evaluates to `()` and returns immediately. The only
place where this seems to be required is the `BufferedPipeReader`
plugin. I'm moving the spawing of the green thread to that plugin
itself so as to avoid this overhead for the other monitors where this
isn't used.

Fixes https://github.com/jaor/xmobar/issues/451